Switches to using getter for returning cell count
Adding a getter allows me to abstract the implementation of the cell
list. My first test expects an integer to be returned. If I start by
creating an array of cells I've had to make assumptions that aren't
covered by tests. I may also have to change a number of tests in future
if I decide that an array (or whatever) isn't the best data structure to
hold my cells in future